# perf(tip5): Avoid one permutation in `hash_varlen`